maximize immunogenicity of epitopes for vaccine
for APMA seminar project. the script that was run was find_epitopes.mat
(data was read in from excel and text files). output of script are
contained in mat files.
x contains the final vector with the binary variables denoting chosen
variables. match refers to the index of the chosen epitopes from the
original 4461 in the candidate set.